refactor: move useKeyboardActionLockerWhileComposing to common hook directory
