+++ packages/bezier-react/src/components/Forms/Inputs/TextField/TextField.test.tsx
@@ -7,9 +7,8 @@ import {
 } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 
+import { COMMON_IME_CONTROL_KEYS } from '~/src/hooks/useKeyboardActionLockerWhileComposing'
 import { render } from '~/src/utils/test'
-
-import { COMMON_IME_CONTROL_KEYS } from '~/src/components/Forms'
 
 import {
   TEXT_INPUT_CLEAR_ICON_TEST_ID,
