--- packages/bezier-react/src/components/Forms/Inputs/constants/CommonImeControlKeys.ts
@@ -1,1 +1,0 @@
-export const COMMON_IME_CONTROL_KEYS = ['Enter', 'Escape', 'Tab', ' ', 'ArrowUp', 'ArrowDown', 'ArrowLeft', 'ArrowRight']
